AIH - Delivery/requirements.txt
Django==3.0.5
pyparsing==2.4.7
whitenoise==5.2.0
gunicorn==20.0.4
xhtml2pdf==0.2.5
dj-database-url==0.5.0
psycopg2-binary==2.8.6
django-filter==2.4.0
reportlab==3.6.6
django-import-export==3.0.1
django-rangefilter==0.9.1
